requirements.txt
Flask
transformers
torch
pillow
requests
timm